requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-jwt-extended==4.5.3
requests==2.31.0
pydantic==2.4.2
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
kafka-python==2.0.2
pika==1.3.2
stomp.py==8.1.0
sqlalchemy==2.0.23
pymongo==4.6.0
psycopg2-binary==2.9.9
mysql-connector-python==8.2.0
cx-Oracle==8.3.0
pyodbc==5.0.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
python-dateutil==2.8.2
jsonschema==4.20.0
pyyaml==6.0.1
cryptography==41.0.8

